requirements.txt
django==3.2.2
argon2-cffi==21.3.0
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
pillow==9.5.0
psycopg2==2.9.6
stripe==2.58.0
boto3==1.17.97
django-storages==1.11.1
django-taggit==1.5.1
django-allauth==0.44.0
django-rest-auth==0.9.5
django-crispy-forms==1.12.0
django-ckeditor==6.1.0
markdown==3.3.4
gunicorn==20.1.0
whitenoise==5.2.0
channels==3.0.4
django-redis==5.0.0
sqlparse==0.3.1
asgiref==3.4.1
pytz==2020.5
mutagen==1.45.1
django-celery-beat==2.2.1
celery==5.1.2
python-dateutil==2.8.1
django-import-export==3.1.0
django-jazzmin==2.5.0
dj_database_url==1.3.0